+++ relvar-storage/src/storage/heap/tests/mod.rs
@@ -1,0 +1,12 @@
+#![allow(unused_imports)]
+#![cfg(test)]
+
+mod common;
+mod corruption;
+mod delete;
+mod gc;
+mod insert;
+mod other;
+mod scan;
+mod update;
+mod version;
